Add some Acaryochloris marina protein data
This will be useful as test data, as well as being in our default
deploy of sequence sets. It satifisies our need for a protein
sequence set. Before this, we had only nucleotide sequence sets.
This will allow testing the filtering of sequence sets based on
which BLAST program is chosen.