Getting closer to generating composite sequence set databases on the fly
The feature of allowing a user to blast against multiple, arbitrary sequence
sets was not clear to me until I hashed it out with rbuels++ recently. The
current vision for implementing this feature is:

1) Modify ExtJS grid to send all selected sequence sets to the form, instead
of only the first
2) When a multi sequence set request comes in, we:
3) Sort them by name alphabetically, Get the metadata about each sequence set
and concatenate all their names together seperated by a hyphen "-"
4) Error out if any of the mimosa_sequence_set_id's we were given are invalid
5) take the sha1 of the string of all the names concatenated, and see if a
cached version of the database exists already, i.e. -e t/data/SHA1.seq
6) If it does, then feed that to Bio::GMOD::Database and we are DONE
7) If not, concatenate all the *.seq files for each sequence set, and save it
in the file t/data/SHA1.seq
8) Hand off to Bio::GMOD::Database

NOTE: This commit does not quite sort the sequence sets correctly.